requirements.txt
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.9.0
attrs==25.3.0
backoff==2.2.1
beautifulsoup4==4.13.3
boto3==1.37.26
botocore==1.37.26
browser-use==0.1.40
cached-property==2.0.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
docopt==0.6.2
fire==0.7.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
install_playwright==0.1.0
jiter==0.9.0
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain-anthropic==0.3.3
langchain-core==0.3.49
langchain-ollama==0.2.2
langchain-openai==0.3.1
langsmith==0.3.19
markdownify==0.14.1
monotonic==1.6
nova-act==1.0.2235.0
numpy==2.2.4
ollama==0.4.7
openai==1.69.0
orjson==3.10.16
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pipreqs==0.4.13
playwright==1.48.0
portkey-ai==1.11.2
posthog==3.23.0
py==1.11.0
pycparser==2.22
pydantic==2.11.1
pydantic_core==2.33.0
pyee==12.0.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
retry==0.9.2
rpds-py==0.24.0
s3transfer==0.11.4
selenium==4.30.0
setuptools==78.1.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
tenacity==9.0.0
termcolor==3.0.1
tiktoken==0.9.0
tqdm==4.67.1
trio==0.29.0
trio-websocket==0.12.2
types-requests==2.32.0.20250328
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.3.0
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
yarg==0.1.10
zstandard==0.23.0
